+++ db/migrate/20200901144806_rm_menu_item_sti.rb
@@ -2,6 +2,6 @@
 
 class RmMenuItemSti < ActiveRecord::Migration[6.0]
   def change
-    remove_column :folio_menu_items, :type
+    remove_column :folio_menu_items, :type, :string
   end
 end